+++ src/ui/stats.js
@@ -38,6 +38,9 @@ export class LibWrapperStats {
 
 		this.PACKAGES  = new Set();
 		this.CONFLICTS = new Map();
+
+		// Seal to prevent accidental modification
+		Object.seal(this);
 	}
 
 	static register_package(package_info) {
@@ -50,34 +53,18 @@ export class LibWrapperStats {
 		this.PACKAGES.add(package_info.key);
 	}
 
-	static register_conflict(package_info, other_info, target) {
+	static register_conflict(package_info, other_info, wrapper, ignored) {
 		if(!this.collect_stats)
 			return;
 
-		if(!other_info)
-			return;
-
-		if(Array.isArray(other_info)) {
-			let notify = false;
-			other_info.forEach((other) => {
-				notify |= LibWrapperStats.register_conflict(package_info, other, target);
-			});
-			return notify;
-		}
-
-		// We first notify everyone that an override was just lost. This hook being handled will prevent us from registering the package conflict
-		if(Hooks.call('libWrapper.ConflictDetected', package_info.id, other_info.id, target) === false) {
-			console.debug(`Conflict between ${package_info.logString} and ${other_info.logString} over '${target}' ignored, as 'libWrapper.ConflictDetected' hook returned false.`);
-			return false;
-		}
-
-		// We now register the conflict
+		// Grab conflict data from storage, or create it if this is a new conflict
 		const key = `${package_info.key}/${other_info.key}`;
 
 		let data = this.CONFLICTS.get(key);
 		if(!data) {
 			data = {
 				count: 0,
+				ignored: 0,
 				package_info: package_info,
 				other_info: other_info,
 				targets: new Map()
@@ -85,11 +72,25 @@ export class LibWrapperStats {
 			this.CONFLICTS.set(key, data);
 		}
 
-		data.count++;
-		data.targets.set(target, (data.targets.get(target) ?? 0) + 1);
+		const target = wrapper.name;
+		let target_data = data.targets.get(target);
+		if(!target_data) {
+			target_data = {
+				count: 0,
+				ignored: 0
+			}
+			data.targets.set(target, target_data);
+		}
 
-		// Done
-		return true;
+		// Increment the conflict counter
+		if(!ignored) {
+			data.count++;
+			target_data.count++;
+		}
+		else {
+			data.ignored++;
+			target_data.ignored++;
+		}
 	}
 
 	static get conflicts() {