+++ src/utils/polyfill.js
@@ -3,88 +3,81 @@
 
 'use strict';
 
+import {LibWrapperInternalError} from './errors/base_errors.js';
 
-// Helper to throw errors
-// This is so we can avoid importing errors.js, to avoid a import loop
-function throw_error(message) {
-    if(globalThis.libWrapper)
-        throw new globalThis.libWrapper.LibWrapperInternalError(message);
-    else
-        throw new Error(message);
-}
 
 
 // game.user.data polyfill, so it can be used before 'ready'
 export const game_user_data = function(return_null=false) {
-    // Try game.user.data first
-    const orig_game_user_data = game?.user?.data;
-    if(orig_game_user_data)
-        return orig_game_user_data;
+	// Try game.user.data first
+	const orig_game_user_data = game?.user?.data;
+	if(orig_game_user_data)
+		return orig_game_user_data;
 
-    // Grab the user ID
-    const userid = game.userId ?? game.data.userId;
-    if(!userid) {
-        if(return_null)
-            return null;
-        throw_error("Unable to obtain the current user ID");
-    }
+	// Grab the user ID
+	const userid = game.userId ?? game.data.userId;
+	if(!userid) {
+		if(return_null)
+			return null;
+		throw new LibWrapperInternalError("Unable to obtain the current user ID");
+	}
 
-    // Find user in game.data.users
-    const user_data = game.data.users.find((x) => { return x._id == userid });
-    if(!user_data) {
-        if(return_null)
-            return null;
-        throw_error("Unable to obtain the current user data object");
-    }
+	// Find user in game.data.users
+	const user_data = game.data.users.find((x) => { return x._id == userid });
+	if(!user_data) {
+		if(return_null)
+			return null;
+		throw new LibWrapperInternalError("Unable to obtain the current user data object");
+	}
 
-    // Done
-    return user_data;
+	// Done
+	return user_data;
 }
 
 // game.user.can polyfill, so it can be used before 'ready'
 export const game_user_can = function(action, return_null=false) {
-    // Try game.user.can first
-    const orig_game_user_can = game?.user?.can;
-    if(orig_game_user_can)
-        return orig_game_user_can(action);
+	// Try game.user.can first
+	const orig_game_user_can = game?.user?.can;
+	if(orig_game_user_can)
+		return orig_game_user_can(action);
 
-    // Obtain game.user.data
-    const user_data = game_user_data(return_null);
-    if(!user_data)
-        return null;
+	// Obtain game.user.data
+	const user_data = game_user_data(return_null);
+	if(!user_data)
+		return null;
 
-    // Check if user is GM
-    if(user_data.role === 4)
-        return true;
+	// Check if user is GM
+	if(user_data.role === 4)
+		return true;
 
-    // Check if the action is in the per-user permissions
-    if(action in user_data.permissions) return user_data.permissions[action];
+	// Check if the action is in the per-user permissions
+	if(action in user_data.permissions) return user_data.permissions[action];
 
-    // Otherwise, check the role's permissions
-    const game_permissions_str = game.data.settings.find((x) => { return x.key === 'core.permissions'});
-    if(game_permissions_str?.value) {
-        const game_permissions = JSON.parse(game_permissions_str.value);
+	// Otherwise, check the role's permissions
+	const game_permissions_str = game.data.settings.find((x) => { return x.key === 'core.permissions'});
+	if(game_permissions_str?.value) {
+		const game_permissions = JSON.parse(game_permissions_str.value);
 
-        const rolePerms = game_permissions[action];
-        if(rolePerms && rolePerms.includes(user_data.role))
-            return true;
-    }
+		const rolePerms = game_permissions[action];
+		if(rolePerms && rolePerms.includes(user_data.role))
+			return true;
+	}
 
-    return false;
+	return false;
 }
 
 // game.user.isGM polyfill, so it can be used before 'ready'
 export const game_user_isGM = function(return_null=false) {
-    // Try game.user.isGM first
-    const orig_game_user_isGM = game?.user?.isGM;
-    if(orig_game_user_isGM !== undefined)
-        return orig_game_user_isGM;
+	// Try game.user.isGM first
+	const orig_game_user_isGM = game?.user?.isGM;
+	if(orig_game_user_isGM !== undefined)
+		return orig_game_user_isGM;
 
-    // Obtain game.user.data
-    const user_data = game_user_data(return_null);
-    if(!user_data)
-        return null;
+	// Obtain game.user.data
+	const user_data = game_user_data(return_null);
+	if(!user_data)
+		return null;
 
-    // Done
-    return user_data.role === 4;
+	// Done
+	return user_data.role === 4;
 }